Skia: Merge upstream patch series that refactors SkCanvas to avoid subclassing.
This change merges upstream Skia revisions r604,r605,r618, and r623. The patch
series refactors SkCanvas so that backends don't need to override it. The r623
patch was upstreamed explicitly so that this series can be made into a clean
patch for Android to avoid merge conflicts in the future.

The log messages from the original commits are as follows:

* r604: http://code.google.com/p/skia/source/detail?r=604
Refactor SkCanvas so that backends don't need to override it.

Methods or classes that should go away are marked deprecated. The only thing I
know of that breaks backward compatibility is SkCanvas((SkDevice*)NULL), but
that is fairly unlikely to occur in the wild because that constructor had a
default value of NULL.

* r605: http://code.google.com/p/skia/source/detail?r=605
Fix a memory leak in the new Canvas/Device workflow.

The previous change made it difficult to inherit from SkCanvas without leaking
memory.  By making SkDeviceFactory not reference counted, the right thing
happens more naturally, just NewCanvas : public SkCanvas(new NewDeviceFactory())
{...}

* r618: http://code.google.com/p/skia/source/detail?r=618
Move the device capability method to SkDevice.

These are not the capabilities of the factory, but of the device.  Additionally,
it is more often needed when you have a device then when you have a factory,
which caused creating of a new factory.

* r623: http://code.google.com/p/skia/source/detail?r=623
Remove include of SkGLDevice.h from SkGLCanvas.h

The '#include "SkGLDevice.h"' from include/core/SkDevice.h requires internal
Skia code to be added to the include search paths when using the deprecated API.
This change adds back SkGLCanvas.cpp to avoid exposing SkGLDevice.h to the
public API.

The change also includes an explicit virtual destructor on SkDeviceFactory to
silence a -Wnon-virtual-dtor warning and allow for -Werror.

Change-Id: I3442e2801c1d8d8c9cb2a20259f15bc870c393dc