Fix issue with SkPaint references being out of date.
The mechanism used for determining whether an SkPaint object is
the same as a cached version was broken. The problem was that Skia
would blow away the generationID in some situations (assignment
and reset), making that ID completely invalid. This would cause the
displayList rendering code to sometimes make the wrong decision, thinking
that an out of date object was actually still valid.

Change-Id: I5e11eb9ac41e5d87d528b99e3781a4d348f76490